Doubly Linked List
Write a program to implement following operations on the doubly linked list.
a. Insert a node at the front of the linked list.
b. Insert a node at the end of the linked list.
c. Delete a last node of the linked list.
d. Delete a node before specified position.